allow package to be used with turborepo
At the moment the package can be used directly by any Next.js application, but if this package is used transitively (by a package which itself is used by the Next.js application) resolving the wasm file fails.

This happens because `new URL(".", import.meta.url)` is invalid. The first argument must be a file, not a directory.